Made contour it's own class
Added the class "Contour" to the utilities package. Added imports for
contour to vision and boilerVision subsystems
