+++ src/org/usfirst/frc/team294/robot/OI.java
@@ -70,7 +70,7 @@ public class OI {//Hopefully this works or maybe this will
 
 	     // Encoders
 //	     SmartDashboard.putData("Drive 10 feet", new DriveWithEncoders(10));
-	     SmartDashboard.putData("Get Boiler Distance", new DisplayBoilerDistance());
+	    // SmartDashboard.putData("Get Boiler Distance", new DisplayBoilerDistance());
 
 
 	     // Stop Command
